+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 调整 API `/api/inbox/getShorthand` | Adjust API `/api/inbox/getShorthand`
 - 调整 API `/api/system/getConf` | Adjust API `/api/system/getConf`
 - 添加 API `/api/filetree/createDailyNote` | Add API `/api/filetree/createDailyNote`
+- 完善 API `/api/system/getConf` | Perfect API `/api/system/getConf`
 
 ## v0.3.5 / 2023-11-28